ray/requirements.txt
ray[serve]==2.35.0
fastapi==0.115.0
vellum-ai==0.14.0
transformers==4.46.0
torch==2.4.1
accelerate==1.0.0
